+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.0] - 2019-6-29
+### Added
+- Get-GraphAzureADUser Function
+- Get-GraphUsersLoggedOn Function
+
+### Changed
+- Get-GraphAzureADGroup Function: API endpoint parameter added. Defaults to beta.
+
 ## [2.3.0] - 2019-4-12
 ### Added
 - Connect-Graph now supports interactive authentication by default to trigger the Enterprise Application creation of "Microsoft Intune PowerShell" and supports the usage of MFA. Non-interactive authentication works as before by passing a PSCredential object to the `Credential` parameter.